+++ packages/applications/ssr/src/app/laureats/[identifiant]/representant-legal/_helpers/steps/index.ts
@@ -1,4 +1,3 @@
 export * from './SaisieNom.step';
 export * from './SaisiePièceJustificative.step';
-export * from './SaisieType.step';
-export * from './SaisieTypeSociété.step';
+export * from './SaisieTypeAndSociété.step';
