+++ packages/applications/ssr/src/app/laureats/[identifiant]/representant-legal/changement/[date]/corriger/CorrigerChangementReprésentantLégal.form.tsx
@@ -11,7 +11,7 @@ import type { ValidationErrors } from '@/utils/formAction';
 import {
   SaisieNomStep,
   SaisiePièceJustificativeStep,
-  SaisieTypeStep,
+  SaisieTypeAndSociétéStep,
   type TypeSociété,
 } from '../../../_helpers/steps';
 import {
@@ -71,8 +71,7 @@ export const CorrigerChangementReprésentantLégalForm: FC<
       />
 
       <div>
-        <SaisieTypeStep
-          contexte="corriger"
+        <SaisieTypeAndSociétéStep
           typeReprésentantLégal={state.typeReprésentantLégal}
           typeSociété={state.typeSociété}
           validationErrors={validationErrors}
